move ProcessWithContext to own file
